Get info about use images from installed proc.yaml
Looking at the Build table in the database is not correct.
There may be multiple build entries for the same (app, tag), using
different "os_image", therefore it may appear as if old "os_image"s
are still used, when in fact they are not.

Instead, just parse all installed "proc.yaml" and get the used
"image_name" from there.